moved round_it() function definition from interface/utils.py to utils/utils.py because it is used in multiple files.
