+++ aviary/utils/propeller_map_conversion.py
@@ -7,11 +7,11 @@ from enum import Enum
 
 import numpy as np
 
-from aviary.interface.utils import round_it
-from aviary.utils.named_values import NamedValues
+from aviary.utils.conversion_utils import _parse, _read_map, _rep
 from aviary.utils.csv_data_file import write_data_file
 from aviary.utils.functions import get_path
-from aviary.utils.conversion_utils import _parse, _read_map, _rep
+from aviary.utils.named_values import NamedValues
+from aviary.utils.utils import round_it
 
 
 class PropMapType(Enum):
